Fix clippy -D warnings across the workspace (dolores CI now green)
All findings on the CI-matching scope (workspace, excluding the
macOS/Windows-only gfx-metal/gfx-dx12/gui-win32 backends dolores's Linux
CI already excludes) are cleared:

- Mechanical/behavior-preserving: derivable Default impls, manual
  is_multiple_of, unnecessary same-type casts, collapsible if-in-match,
  field-assignment-outside-initializer, explicit into_iter, io::Error::other,
  manual C string literals, manual slice size calc, chunks_exact ->
  as_chunks, useless .into() conversions, netbsd-only bin tools' unused
  imports/dead code now properly cfg-gated instead of unconditional,
  items-after-test-module reordered in camera_preview.rs.
- text_document::TextDocument's inherent to_string() became a real
  Display impl instead - .to_string() still works identically for every
  caller via the blanket ToString impl, zero API change.
- too_many_arguments: #[allow]'d with a justifying comment on private
  single-call-site helpers and on two genuinely public, externally-consumed
  signatures (render_text_lines - called directly by sng-rusty - and
  linux_egl::present_scene) where restructuring would be a real breaking
  API change out of scope for a lint pass.
- host-desktop/src/linux.rs's append_rasterized_polyline_images (and its
  only caller) is real, implemented-but-not-yet-wired-up logic, not
  throwaway dead code - preserved with #[allow(dead_code)] rather than
  deleted.

Verified with the actual scope dolores's CI checks (aarch64-unknown-linux-gnu,
run natively on dolores itself, not cross-compiled) plus the full local
workspace build/test on macOS. Confirmed zero regression across every
known consumer: sng-roguelite (104+32+94 tests), sng-zhoenus (17 tests),
sng-rusty (111 tests, its 2 pre-existing voice-asset failures unchanged
and unrelated), and entitlement-achievement-blockchain, all matching their
pre-change baselines exactly. One pre-existing, unrelated finding
confirmed NOT caused by this change (bisected against the prior commit):
loadngo-gfx-gles's own test target fails to compile on macOS specifically
(a cfg-gated helper only compiled for android/linux) - untouched here,
separate follow-up.

+++ proactor/src/kqueue.rs
@@ -146,13 +146,13 @@ impl CompletionPort for KqueuePort {
         }
 
         if event.filter == EVFILT_USER {
-            if event.ident as usize == QUEUE_IDENT {
+            if event.ident == QUEUE_IDENT {
                 if let Some(envelope) = self.drain_completion() {
                     return Ok(PollEvent::Completion(envelope));
                 }
                 return Ok(PollEvent::Wake);
             }
-            if event.ident as usize == WAKE_IDENT {
+            if event.ident == WAKE_IDENT {
                 return Ok(PollEvent::Wake);
             }
         }
